Update index.md
This looks really good. I made a few tiny changes, including moving the author bio to the end. If there is a link we could add to the "Road to Kubernetes Production (in?) Practice" book, referenced at the end, please add the URL.  In general, it's easier for me to review content in Google Docs than in Github, if that's possible. I hope I did this correctly here, it has been a while.